Move postNewMessage to a store action
Removes the store usage in messagesService
