+++ src/components/NewMessageForm/NewMessageForm.vue
@@ -116,8 +116,7 @@
 
 <script>
 import AdvancedInput from './AdvancedInput/AdvancedInput'
-import { getFilePickerBuilder, showError } from '@nextcloud/dialogs'
-import { postNewMessage } from '../../services/messagesService'
+import { getFilePickerBuilder } from '@nextcloud/dialogs'
 import { getCapabilities } from '@nextcloud/capabilities'
 import Quote from '../Quote'
 import Actions from '@nextcloud/vue/dist/Components/Actions'
@@ -128,7 +127,6 @@ import { shareFile } from '../../services/filesSharingServices'
 import { CONVERSATION } from '../../constants'
 import EmoticonOutline from 'vue-material-design-icons/EmoticonOutline'
 import Send from 'vue-material-design-icons/Send'
-import CancelableRequest from '../../utils/cancelableRequest'
 
 const picker = getFilePickerBuilder(t('spreed', 'File to share'))
 	.setMultiSelect(false)
@@ -322,76 +320,7 @@ export default {
 				EventBus.$emit('smoothScrollChatToBottom')
 				// Also remove the message to be replied for this conversation
 				this.$store.dispatch('removeMessageToBeReplied', this.token)
-				let timeout
-				try {
-					// Posts the message to the server
-					const { request, cancel } = CancelableRequest(postNewMessage)
-
-					timeout = setTimeout(() => {
-						cancel('canceled')
-						this.$store.dispatch('markTemporaryMessageAsFailed', {
-							message: temporaryMessage,
-							reason: 'timeout',
-						})
-					}, 30000)
-					const response = await request(temporaryMessage)
-					clearTimeout(timeout)
-
-					// If successful, deletes the temporary message from the store
-					this.$store.dispatch('removeTemporaryMessageFromStore', temporaryMessage)
-
-					const message = response.data.ocs.data
-					// And adds the complete version of the message received
-					// by the server
-					this.$store.dispatch('processMessage', message)
-
-					// update lastMessage and lastReadMessage
-					// do it conditionally because there could have been more messages appearing concurrently
-					if (this.conversation.lastMessage && message.id > this.conversation.lastMessage.id) {
-						this.$store.dispatch('updateConversationLastMessage', {
-							token: this.token,
-							lastMessage: message,
-						})
-					}
-					if (message.id > this.conversation.lastReadMessage) {
-						this.$store.dispatch('updateLastReadMessage', {
-							token: this.token,
-							id: message.id,
-							updateVisually: true,
-						})
-					}
-				} catch (error) {
-					let statusCode = null
-					console.debug(`error while submitting message ${error}`, error)
-					if (error.isAxiosError) {
-						statusCode = error?.response?.status
-					}
-
-					if (timeout) {
-						clearTimeout(timeout)
-					}
-
-					// 403 when room is read-only, 412 when switched to lobby mode
-					if (statusCode === 403) {
-						showError(t('spreed', 'No permission to post messages in this conversation'))
-						this.$store.dispatch('markTemporaryMessageAsFailed', {
-							message: temporaryMessage,
-							reason: 'read-only',
-						})
-					} else if (statusCode === 412) {
-						showError(t('spreed', 'No permission to post messages in this conversation'))
-						this.$store.dispatch('markTemporaryMessageAsFailed', {
-							message: temporaryMessage,
-							reason: 'lobby',
-						})
-					} else {
-						showError(t('spreed', 'Could not post message: {errorMessage}', { errorMessage: error.message || error }))
-						this.$store.dispatch('markTemporaryMessageAsFailed', {
-							message: temporaryMessage,
-							reason: 'other',
-						})
-					}
-				}
+				await this.$store.dispatch('postNewMessage', temporaryMessage)
 			}
 		},
